+++ src/realm/array.hpp
@@ -81,6 +81,11 @@ template<class T> inline T no0(T v) { return v == 0 ? 1 : v; }
 /// found'. It is similar in function to std::string::npos.
 const std::size_t npos = std::size_t(-1);
 
+// Represents null in Query, find(), get(), set(), etc.
+struct null {
+    operator StringData() { return StringData(); }
+};
+
 /// Alias for realm::npos.
 const std::size_t not_found = npos;
 
@@ -815,6 +820,7 @@ public:
 
     template<class TreeTraits> struct TreeInsert: TreeInsertBase {
         typename TreeTraits::value_type m_value;
+        bool m_nullable;
     };
 
     /// Same as bptree_insert() but insert after the last element.
@@ -839,8 +845,10 @@ public:
 
     /// Like get(const char*, std::size_t) but gets two consecutive
     /// elements.
-    static std::pair<int_least64_t, int_least64_t> get_two(const char* header,
+    static std::pair<int64_t, int64_t> get_two(const char* header,
                                                            std::size_t ndx) REALM_NOEXCEPT;
+
+    static void get_three(const char* data, size_t ndx, ref_type& v0, ref_type& v1, ref_type& v2) REALM_NOEXCEPT;
 
     /// The meaning of 'width' depends on the context in which this
     /// array is used.
@@ -928,8 +936,9 @@ protected:
 
     bool do_erase_bptree_elem(std::size_t elem_ndx, EraseHandler&);
 
-
-    template <IndexMethod method, class T> size_t index_string(StringData value, Column& result, size_t &result_ref, void* column, StringGetter get_func) const;
+    template <IndexMethod method, class T> 
+    size_t index_string(StringData value, Column& result, size_t &result_ref, void* column, 
+                        StringGetter get_func) const;
 protected:
 //    void AddPositiveLocal(int64_t value);