+++ src/realm/array_string.hpp
@@ -24,14 +24,33 @@
 
 namespace realm {
 
+/*
+ArrayString stores strings as a concecutive list of fixed-length blocks of m_width bytes. The 
+longest string it can store is (m_width - 1) bytes before it needs to expand.
+
+An example of the format for m_width = 4 is following sequence of bytes, where x is payload:
+
+xxx0 xx01 x002 0003 0004 (strings "xxx",. "xx", "x", "", realm::null())
+
+So each string is 0 terminated, and the last byte in a block tells how many 0s are present, except
+for a realm::null() which has the byte set to m_width (4). The byte is used to compute the length of a string
+in various functions.
+
+New: If m_witdh = 0, then all elements are realm::null(). So to add an empty string we must expand m_width
+New: StringData is null() if-and-only-if StringData::data() == 0.
+*/
+
 class ArrayString: public Array {
 public:
     typedef StringData value_type;
-
-    explicit ArrayString(Allocator&) REALM_NOEXCEPT;
+    // Constructor defaults to non-nullable because we use non-nullable ArrayString so many places internally in core
+    // (data which isn't user payload) where null isn't needed.
+    explicit ArrayString(Allocator&, bool nullable = false) REALM_NOEXCEPT;
     explicit ArrayString(no_prealloc_tag) REALM_NOEXCEPT;
     ~ArrayString() REALM_NOEXCEPT override {}
 
+    bool is_null(size_t ndx) const;
+    void set_null(size_t ndx);
     StringData get(std::size_t ndx) const REALM_NOEXCEPT;
     void add();
     void add(StringData value);
@@ -53,7 +72,7 @@ public:
     /// array instance. If an array instance is already available, or
     /// you need to get multiple values, then this method will be
     /// slower.
-    static StringData get(const char* header, std::size_t ndx) REALM_NOEXCEPT;
+    static StringData get(const char* header, std::size_t ndx, bool nullable) REALM_NOEXCEPT;
 
     ref_type bptree_leaf_insert(std::size_t ndx, StringData, TreeInsertBase& state);
 
@@ -84,6 +103,8 @@ private:
     std::size_t CalcItemCount(std::size_t bytes,
                               std::size_t width) const REALM_NOEXCEPT override;
     WidthType GetWidthType() const override { return wtype_Multiply; }
+
+    bool m_nullable;
 };
 
 
@@ -91,8 +112,8 @@ private:
 // Implementation:
 
 // Creates new array (but invalid, call init_from_ref() to init)
-inline ArrayString::ArrayString(Allocator& alloc) REALM_NOEXCEPT:
-    Array(alloc)
+inline ArrayString::ArrayString(Allocator& alloc, bool nullable) REALM_NOEXCEPT:
+Array(alloc), m_nullable(nullable)
 {
 }
 
@@ -122,30 +143,43 @@ inline StringData ArrayString::get(std::size_t ndx) const REALM_NOEXCEPT
 {
     REALM_ASSERT_3(ndx, <, m_size);
     if (m_width == 0)
-        return StringData("", 0);
+        return m_nullable ? realm::null() : StringData("");
+
     const char* data = m_data + (ndx * m_width);
     std::size_t size = (m_width-1) - data[m_width-1];
+
+    if (size == static_cast<size_t>(-1))
+        return m_nullable ? realm::null() : StringData("");
+
+    REALM_ASSERT(data[size] == 0); // Realm guarantees 0 terminated return strings
     return StringData(data, size);
 }
 
 inline void ArrayString::add(StringData value)
 {
+    REALM_ASSERT(!(!m_nullable && value.is_null()));
     insert(m_size, value); // Throws
 }
 
 inline void ArrayString::add()
 {
-    add(StringData()); // Throws
+    add(m_nullable ? realm::null() : StringData("")); // Throws
 }
 
-inline StringData ArrayString::get(const char* header, std::size_t ndx) REALM_NOEXCEPT
+inline StringData ArrayString::get(const char* header, std::size_t ndx, bool nullable) REALM_NOEXCEPT
 {
-    REALM_ASSERT_3(ndx, <, get_size_from_header(header));
+    REALM_ASSERT(ndx < get_size_from_header(header));
     std::size_t width = get_width_from_header(header);
-    if (width == 0)
-        return StringData("", 0);
     const char* data = get_data_from_header(header) + (ndx * width);
+
+    if (width == 0)
+        return nullable ? realm::null() : StringData("");
+
     std::size_t size = (width-1) - data[width-1];
+
+    if (size == static_cast<size_t>(-1))
+        return nullable ? realm::null() : StringData("");
+
     return StringData(data, size);
 }