+++ src/realm/column_binary.cpp
@@ -41,7 +41,8 @@ ColumnBinary::ColumnBinary(Allocator& alloc, ref_type ref)
             return;
         }
         // Big blobs root leaf
-        ArrayBigBlobs* root = new ArrayBigBlobs(alloc); // Throws
+        // fixme, modify the 'nullable' arguments to constructor to support null for binary columns
+        ArrayBigBlobs* root = new ArrayBigBlobs(alloc, false); // Throws
         root->init_from_mem(mem);
         m_array.reset(root);
         return;
@@ -66,7 +67,7 @@ struct SetLeafElem: Array::UpdateHandler {
     {
         bool is_big = Array::get_context_flag_from_header(mem.m_addr);
         if (is_big) {
-            ArrayBigBlobs leaf(m_alloc);
+            ArrayBigBlobs leaf(m_alloc, false);
             leaf.init_from_mem(mem);
             leaf.set_parent(parent, ndx_in_parent);
             leaf.set(elem_ndx_in_leaf, m_value, m_add_zero_term); // Throws
@@ -80,7 +81,7 @@ struct SetLeafElem: Array::UpdateHandler {
             return;
         }
         // Upgrade leaf from small to big blobs
-        ArrayBigBlobs new_leaf(m_alloc);
+        ArrayBigBlobs new_leaf(m_alloc, false);
         new_leaf.create(); // Throws
         new_leaf.set_parent(parent, ndx_in_parent); // Throws
         new_leaf.update_parent(); // Throws
@@ -180,7 +181,7 @@ ref_type ColumnBinary::leaf_insert(MemRef leaf_mem, ArrayParent& parent,
     InsertState& state_2 = static_cast<InsertState&>(state);
     bool is_big = Array::get_context_flag_from_header(leaf_mem.m_addr);
     if (is_big) {
-        ArrayBigBlobs leaf(alloc);
+        ArrayBigBlobs leaf(alloc, false);
         leaf.init_from_mem(leaf_mem);
         leaf.set_parent(&parent, ndx_in_parent);
         return leaf.bptree_leaf_insert(insert_ndx, state_2.m_value, state_2.m_add_zero_term,
@@ -193,7 +194,7 @@ ref_type ColumnBinary::leaf_insert(MemRef leaf_mem, ArrayParent& parent,
         return leaf.bptree_leaf_insert(insert_ndx, state_2.m_value, state_2.m_add_zero_term,
                                        state); // Throws
     // Upgrade leaf from small to big blobs
-    ArrayBigBlobs new_leaf(alloc);
+    ArrayBigBlobs new_leaf(alloc, false);
     new_leaf.create(); // Throws
     new_leaf.set_parent(&parent, ndx_in_parent);
     new_leaf.update_parent(); // Throws
@@ -229,7 +230,7 @@ public:
             return false;
         }
         // Big blobs
-        ArrayBigBlobs leaf(get_alloc());
+        ArrayBigBlobs leaf(get_alloc(), false);
         leaf.init_from_mem(leaf_mem);
         leaf.set_parent(parent, leaf_ndx_in_parent);
         REALM_ASSERT_3(leaf.size(), >=, 1);
@@ -258,7 +259,7 @@ public:
         }
         else {
             // Big blobs
-            ArrayBigBlobs* leaf_2 = new ArrayBigBlobs(get_alloc()); // Throws
+            ArrayBigBlobs* leaf_2 = new ArrayBigBlobs(get_alloc(), false); // Throws
             leaf_2->init_from_mem(leaf_mem);
             leaf = leaf_2;
         }
@@ -324,9 +325,7 @@ void ColumnBinary::do_move_last_over(size_t row_ndx, size_t last_row_ndx)
     std::unique_ptr<char[]> buffer(new char[value.size()]); // Throws
     copy(value.data(), value.data()+value.size(), buffer.get());
     BinaryData copy_of_value(buffer.get(), value.size());
-
     set(row_ndx, copy_of_value); // Throws
-
     bool is_last = true;
     erase(last_row_ndx, is_last); // Throws
 }
@@ -377,7 +376,7 @@ bool ColumnBinary::upgrade_root_leaf(size_t value_size)
     ArrayBinary* leaf = static_cast<ArrayBinary*>(m_array.get());
     Allocator& alloc = leaf->get_alloc();
     std::unique_ptr<ArrayBigBlobs> new_leaf;
-    new_leaf.reset(new ArrayBigBlobs(alloc)); // Throws
+    new_leaf.reset(new ArrayBigBlobs(alloc, false)); // Throws
     new_leaf->create(); // Throws
     new_leaf->set_parent(leaf->get_parent(), leaf->get_ndx_in_parent());
     new_leaf->update_parent(); // Throws
@@ -421,7 +420,7 @@ public:
             return leaf.slice(offset, size, target_alloc); // Throws
         }
         // Big blobs
-        ArrayBigBlobs leaf(m_alloc);
+        ArrayBigBlobs leaf(m_alloc, false);
         leaf.init_from_mem(leaf_mem);
         return leaf.slice(offset, size, target_alloc); // Throws
     }
@@ -511,7 +510,7 @@ void ColumnBinary::refresh_accessor_tree(size_t, const Spec&)
         }
         else {
             // New root is 'big blobs' leaf
-            ArrayBigBlobs* root = new ArrayBigBlobs(alloc); // Throws
+            ArrayBigBlobs* root = new ArrayBigBlobs(alloc, false); // Throws
             root->init_from_mem(root_mem);
             new_root = root;
         }
@@ -544,7 +543,7 @@ size_t verify_leaf(MemRef mem, Allocator& alloc)
         return leaf.size();
     }
     // Big blobs
-    ArrayBigBlobs leaf(alloc);
+    ArrayBigBlobs leaf(alloc, false);
     leaf.init_from_mem(mem);
     leaf.Verify();
     return leaf.size();
@@ -598,7 +597,7 @@ void ColumnBinary::leaf_to_dot(MemRef leaf_mem, ArrayParent* parent, size_t ndx_
         return;
     }
     // Big blobs
-    ArrayBigBlobs leaf(m_array->get_alloc());
+    ArrayBigBlobs leaf(m_array->get_alloc(), false); // fixme, null support for to_dot
     leaf.init_from_mem(leaf_mem);
     leaf.set_parent(parent, ndx_in_parent);
     leaf.to_dot(out, is_strings);
@@ -621,7 +620,7 @@ void leaf_dumper(MemRef mem, Allocator& alloc, ostream& out, int level)
     }
     else {
         // Big blobs
-        ArrayBigBlobs leaf(alloc);
+        ArrayBigBlobs leaf(alloc, false); // fixme, nulls not yet supported by leaf dumper
         leaf.init_from_mem(mem);
         leaf_size = leaf.size();
         leaf_type = "Big blobs leaf";