+++ src/realm/util/file.cpp
@@ -917,6 +917,29 @@ void File::move(const string& old_path, const string& new_path)
     }
 }
 
+bool File::copy(string source, string destination)
+{
+    // Quick and dirty file copy, only used for unit tests. Todo, make more robust if used by Core.
+    char buf[1024];
+    size_t read;
+    File::try_remove(destination);
+    FILE* src = fopen(source.c_str(), "rb");
+    if (!src)
+        return false;
+
+    FILE* dst = fopen(destination.c_str(), "wb");
+    if (!dst) {
+        fclose(src);
+        return false;
+    }
+
+    while ((read = fread(buf, 1, 1024, src))) {
+        fwrite(buf, 1, read, dst);
+    }
+    fclose(src);
+    fclose(dst);
+    return true;
+}
 
 bool File::is_same_file(const File& f) const
 {