+++ test/test_file.cpp
@@ -214,6 +214,8 @@ TEST(File_SetEncryptionKey)
 #endif
 }
 
+#ifndef _WIN32
+
 TEST(File_ReadWrite)
 {
     TEST_PATH(path);
@@ -230,6 +232,8 @@ TEST(File_ReadWrite)
         CHECK_EQUAL(i, read);
     }
 }
+
+#endif
 
 TEST(File_Resize)
 {