+++ release_notes.md
@@ -11,11 +11,10 @@
 
 ### API breaking changes:
 
-* Lorem ipsum.
 
 ### Enhancements:
 
-* Improved performance of advance_read() over commits with string or binary data insertions.
+* Null support for String columns. Call add_column() with nullable = true. You can then use tightdb::null() in place of any StringData (in Query, Table::find(), get(), set(), etc) for that column. You can also call Table::is_null(), Table::set_null(), StringData::is_null() and Table::is_nullable(size_t col_ndx).
 * Improved performance sorting TableView and LinkView.
 * Added Table::remove_search_index().